// actions/update-card-order/schema.ts
import * as z from "zod";

export const UpdateCardOrder = z.object({
    items: z.array(
        z.object({
            id: z.string(),
            order: z.number(),
            title: z.string(),
            listId: z.string(),
            createdAt: z.date(),
            updatedAt: z.date()
        })
    ),
    boardId: z.string()
});